Updated validator API of cef and interfaces concept - (Task #113)
