+++ de.dlr.sc.virsat.model.extension.cef.interfaces.test/src-gen/de/dlr/sc/virsat/model/extension/cef/interfaces/validator/AInterfacesValidatorTest.java
@@ -1,0 +1,43 @@
+/*******************************************************************************
+ * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package de.dlr.sc.virsat.model.extension.cef.interfaces.validator;
+
+// *****************************************************************
+// * Import Statements
+// *****************************************************************
+
+
+import de.dlr.sc.virsat.model.dvlm.structural.StructuralElementInstance;
+import de.dlr.sc.virsat.model.dvlm.structural.StructuralFactory;
+
+
+// *****************************************************************
+// * Class Declaration
+// *****************************************************************
+
+import org.junit.Before;
+
+/**
+ * Auto Generated Abstract Generator Gap Class
+ * 
+ * Don't Manually modify this class
+ * 
+ * Concept for Functional Interface Architecture in CEF
+ * 
+ */	
+public abstract class AInterfacesValidatorTest {
+
+	StructuralElementInstance testSei;
+
+	@Before
+	public void setup() {
+		testSei = StructuralFactory.eINSTANCE.createStructuralElementInstance();
+	}
+}
