--- de.dlr.sc.virsat.model.extension.cef.interfaces/src/de/dlr/sc/virsat/model/extension/cef/interfaces/validator/StructuralElementInstanceValidator.java
@@ -1,34 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *******************************************************************************/
-package de.dlr.sc.virsat.model.extension.cef.interfaces.validator;
-
-import de.dlr.sc.virsat.model.dvlm.structural.StructuralElementInstance;
-import de.dlr.sc.virsat.model.dvlm.validator.IStructuralElementInstanceValidator;
-
-
-// *****************************************************************
-// * Class Declaration
-// *****************************************************************
-
-/**
- * Auto Generated Class inheriting from Generator Gap Class
- *
- * This class is generated once, do your changes here
- *
- * Example Concept for Interface Architecture
- *
- */
-public class StructuralElementInstanceValidator extends AStructuralElementInstanceValidator implements IStructuralElementInstanceValidator {
-
-	@Override
-	public boolean validate(StructuralElementInstance sei) {
-		return super.validate(sei);
-	}
-}
