+++ de.dlr.sc.virsat.model.extension.cef.test/src-gen/de/dlr/sc/virsat/model/extension/cef/validator/ACefValidatorTest.java
@@ -1,0 +1,43 @@
+/*******************************************************************************
+ * Copyright (c) 2008-2019 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
+package de.dlr.sc.virsat.model.extension.cef.validator;
+
+// *****************************************************************
+// * Import Statements
+// *****************************************************************
+
+
+import de.dlr.sc.virsat.model.dvlm.structural.StructuralElementInstance;
+import de.dlr.sc.virsat.model.dvlm.structural.StructuralFactory;
+
+
+// *****************************************************************
+// * Class Declaration
+// *****************************************************************
+
+import org.junit.Before;
+
+/**
+ * Auto Generated Abstract Generator Gap Class
+ * 
+ * Don't Manually modify this class
+ * 
+ * VirSat DLR CEF Concept
+ * 
+ */	
+public abstract class ACefValidatorTest {
+
+	StructuralElementInstance testSei;
+
+	@Before
+	public void setup() {
+		testSei = StructuralFactory.eINSTANCE.createStructuralElementInstance();
+	}
+}
